Updated list contacts
As of this commit, I believe the program is robust to any changes to the fields (see fields.py) between program executions. The only necessary fields to keep are first_name and last_name. All other fields may be removed, and new ones may be added. In the case of new fields added, there are no checks for those not already implemented.